backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
google-cloud-logging==3.9.0
python-multipart==0.0.6
pydantic==2.5.3
starlette==0.36.1
httpx==0.26.0
